moved observations object to predict.py
